Move SATP setup so that it only happens on U54s, not E51 (where it might trap) Merge in latest DDR/mpfs code and latest drivers Merge in platform config for icicle-kit-mtd-demo-design Fix issue with section placement (introduced by switch_code merge), disable PDMA for now Default UART to 115200 (rather than 57600)

+++ baremetal/polarfire-soc-bare-metal-library/src/platform/hal/cpu_types.h
@@ -1,16 +1,12 @@
 /*******************************************************************************
- * Copyright 2019 Microchip Corporation.
+ * Copyright 2019-2020 Microchip FPGA Embedded Systems Solutions.
  *
  * SPDX-License-Identifier: MIT
  *
  * MPFS HAL Embedded Software
  *
  */
-/*******************************************************************************
- * 
- * SVN $Revision$
- * SVN $Date$
- */
+
 #ifndef CPU_TYPES_H
 #define CPU_TYPES_H
 
@@ -25,14 +21,14 @@ extern "C" {
  * Used to specify the address of peripherals present in the processor's memory
  * map.
  */
-typedef unsigned int addr_t;
+typedef unsigned long addr_t;
 
 /*------------------------------------------------------------------------------
  * psr_t: processor state register.
  * Used by HAL_disable_interrupts() and HAL_restore_interrupts() to store the
  * processor's state between disabling and restoring interrupts.
  */
-typedef unsigned int psr_t;
+typedef unsigned long psr_t;
 
 #ifdef __cplusplus
 }
